Make `extractLeadingCheckbox` recursive
The main goal here is to make the logic here easier to read and follow. It also makes it easier to update to handle more complex scenarios if desired in the future. This does make the code slightly more vulnerable to getting slowed down by weird trees (that aren't valid HTML) with `<p>` elements deeply nested inside each other.